Refactored app into multiple files
